% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_readValidationData.R
\name{GEOtop_ReadValidationData}
\alias{GEOtop_ReadValidationData}
\title{Read point output from GEOtop for verification of the model results}
\usage{
GEOtop_ReadValidationData(wpath, obs, lookup_tbl_observation = NULL,
  soil_files = TRUE, save_rData = TRUE, tz = "Etc/GMT-1", level = 1)
}
\arguments{
\item{wpath}{path into simulation folder}

\item{obs}{zoo object, data frame, with specific names of variables used for validate the model results; name conventions according to CF Standard Name Table \url{http://cfconventions.org}.}

\item{soil_files}{boolean, \code{TRUE}: soil files are provided as GEOtop input. \code{FALSE}: soil is parameterized in the geotop.inpts file}

\item{save_rData}{boolean, if \code{TRUE} (default) data is stored in working directory (simulation folder)}
}
\description{
Read point output from GEOtop for verification of the model results
}
\examples{
 ## TO DO 
wpath <- '/home/ecor/activity/2016/eurac2016/Incarico_EURAC/Simulations/B2/B2_BeG_017_DVM_001' 
load(file.path(wpath, "obs", "observation.RData"))
out <- GEOtop_ReadValidationData(wpath = wpath, obs = observation, save_rData = TRUE)



}
\author{
Johannes Brenner
}

