% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractGeotopVar}
\alias{extractGeotopVar}
\title{Extract variable from an object returend by \code{\link{GEOtop_ReadValidationData}}}
\usage{
extractGeotopVar(x, InputVar = NULL, Add_InputVar = NULL,
  aggregate = c("hourly", "daily", "monthly", "yearly"),
  aggregate_fun = c("mean", "sum"), ...)
}
\arguments{
\item{x}{objectect returned by \code{\link{GEOtop_ReadValidationData}} with \code{merge==TRUE}}

\item{InputVar}{discover variable}

\item{Add_InputVar}{additional variable (simulated)}

\item{aggregate}{aggregate option. Default it is \code{c("hourly","daily","monthly","yearly")} and it is considered the first element.}

\item{aggregate_fun}{aggregate function for \code{InputVar} and \code{Add_InputVar} respectively. Deafault is \code{c("mean","sum")}.}

\item{...}{further arguments}
}
\description{
Extract variable from an object returend by \code{\link{GEOtop_ReadValidationData}}
}
\note{
It is assumed that \code{x} already contains hourly aggregated time series, so in case \code{aggregare=="hourly"} no aggragetion is calculated.
}
\examples{

wpath <- '/home/ecor/activity/2016/eurac2016/idra/B2_BeG_017_DVM_001_test_1' 


ex <- GEOtop_ReadValidationData(wpath = wpath)
 
eout <- extractGeotopVar(ex)
dout <- extractGeotopVar(ex,aggregate="daily")
mout <-  extractGeotopVar(ex,aggregate="monthly")



}

